+++ table/block_based_table_factory.h
@@ -59,5 +59,7 @@ class BlockBasedTableFactory : public TableFactory {
 
 extern const std::string kHashIndexPrefixesBlock;
 extern const std::string kHashIndexPrefixesMetadataBlock;
+extern const std::string kPropTrue;
+extern const std::string kPropFalse;
 
 }  // namespace rocksdb